Remove deprecated Cardano transaction broadcasting logic and refactor UTXO transaction broadcasting to enhance clarity and maintainability. Updated `broadcastUtxoTx` to directly handle transaction broadcasting and error management, improving overall response processing.
